+++ Sources/ASCKit/AscEndpoint.swift
@@ -21,6 +21,12 @@ enum AscEndpoint {
     case read(url: URL, filters: [Filter], limit: UInt?)
 
     case listAppStoreVersions(appId: String, filters: [Filter], limit: UInt?)
+
+    case listAppInfos(appId: String, limit: UInt?)
+
+    case getAgeRatings(appInfoId: String)
+    case updateAgeRatings(ageRatingDeclarationId: String, parameters: [String: Any])
+
     case listAllBetaGroupsForTester(id: String, filters: [Filter], limit: UInt?)
 
     case listAccessibilityDeclarations(appId: String, filters: [Filter], limit: UInt?)
@@ -167,6 +173,14 @@ extension AscEndpoint: Endpoint {
         case let .listAppStoreVersions(appId, _, _):
             "/\(apiVersion)/apps/\(appId)/appStoreVersions"
 
+        case let .listAppInfos(appId, limit):
+            "/\(apiVersion)/apps/\(appId)/appInfos"
+
+        case let .getAgeRatings(appInfoId):
+            "/\(apiVersion)/appInfos/\(appInfoId)/ageRatingDeclaration"
+        case let .updateAgeRatings(ageRatingDeclarationId, _):
+            "/\(apiVersion)/ageRatingDeclarations/\(ageRatingDeclarationId)"
+
         case let .listAccessibilityDeclarations(appId, _, _):
             "/\(apiVersion)/apps/\(appId)/accessibilityDeclarations"
         case .createAccessibilityDeclaration:
@@ -201,8 +215,14 @@ extension AscEndpoint: Endpoint {
              .listAllBetaGroupsForTester(_, let filters, let limit):
             return queryItems(from: filters, limit: limit)
 
+        case .listAppInfos(_, let limit):
+            return queryItems(from: [], limit: limit)
+
         case .inviteBetaTester,
              .addBetaTester,
+
+             .getAgeRatings,
+             .updateAgeRatings,
 
              .createAccessibilityDeclaration,
              .updateAccessibilityDeclaration,
@@ -210,6 +230,7 @@ extension AscEndpoint: Endpoint {
              .publishAccessibilityDeclaration,
 
              .registerBundleId,
+
              .expireBuild:
             return []
         }
@@ -219,6 +240,8 @@ extension AscEndpoint: Endpoint {
         switch self {
         case .read,
              .listAppStoreVersions,
+             .listAppInfos,
+             .getAgeRatings,
              .listAccessibilityDeclarations,
              .listAllBetaGroupsForTester:
             .get
@@ -229,6 +252,7 @@ extension AscEndpoint: Endpoint {
             .post
         case .expireBuild,
              .updateAccessibilityDeclaration,
+             .updateAgeRatings,
              .publishAccessibilityDeclaration:
             .patch
         case .deleteAccessibilityDeclaration:
@@ -264,6 +288,8 @@ extension AscEndpoint: Endpoint {
         switch self {
         case .read,
              .listAppStoreVersions,
+             .listAppInfos,
+             .getAgeRatings,
              .listAccessibilityDeclarations,
              .listAllBetaGroupsForTester,
              .deleteAccessibilityDeclaration:
@@ -311,11 +337,20 @@ extension AscEndpoint: Endpoint {
                 ]
             ]
 
-        case let .updateAccessibilityDeclaration(id, parameters):
+        case let .updateAccessibilityDeclaration(accessibilityDeclarationId, parameters):
             return [
                 "data": [
                     "type": "accessibilityDeclarations",
-                    "id": "\(id)",
+                    "id": "\(accessibilityDeclarationId)",
+                    "attributes": parameters,
+                ]
+            ]
+
+        case let .updateAgeRatings(ageRatingDeclarationId, parameters):
+            return [
+                "data": [
+                    "type": "ageRatingDeclarations",
+                    "id": "\(ageRatingDeclarationId)",
                     "attributes": parameters,
                 ]
             ]